backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
PyMySQL>=1.1
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
bcrypt>=4.0
email-validator>=2.0
python-multipart>=0.0.9
segno>=1.6
openpyxl>=3.1
reportlab>=4.1
python-docx>=1.1
alembic>=1.13
